refactor: move camera related files
