+++ sources/camera/autocamera.cpp
