+++ sources/camera/autocamera.hpp
