+++ sources/camera/camera.cpp
