+++ sources/camera/camera.hpp
