+++ sources/camera/camerainstruction.cpp
