+++ sources/camera/camerainstruction.hpp
